endedAtTime
